address coderabbit comments
