+++ app/packages/state/src/hooks/useTooltip.ts
@@ -1,6 +1,9 @@
 import * as fos from "@fiftyone/state";
 import { useCallback } from "react";
 import { useSetRecoilState } from "recoil";
+import { computeCoordinates } from "./useTooltip.utils";
+
+export type { ComputeCoordinatesReturnType } from "./useTooltip.utils";
 
 export default function useTooltip() {
   const setTooltipCoordinates = useSetRecoilState(fos.tooltipCoordinates);
@@ -14,30 +17,3 @@ export default function useTooltip() {
     setCoords,
   };
 }
-
-type placement = number | "unset";
-
-export function computeCoordinates([x, y]: [number, number]): {
-  bottom?: placement;
-  top?: placement;
-  left?: placement;
-  right?: placement;
-} {
-  let top: placement = y,
-    bottom: placement = "unset";
-  if (y > window.innerHeight / 2) {
-    bottom = window.innerHeight - y;
-    top = "unset";
-  }
-
-  return {
-    bottom,
-    top,
-    left: x <= window.innerWidth / 2 ? x + 24 : "unset",
-    right: x > window.innerWidth / 2 ? window.innerWidth - x + 24 : "unset",
-  };
-}
-
-export type ComputeCoordinatesReturnType = ReturnType<
-  typeof computeCoordinates
->;
